pyproject.toml
[tool.poetry]
name = "frontend"
version = "4.0.0"
description = "georgia frontend"
authors = ["marcelo amorelli <marcelo.amorelli@upetrading.com>"]
readme = "README.md"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.poetry.scripts]
run-main-frontend = "frontend.app:server"

[tool.poetry.dependencies]
python = "~3.11"
dash = "^2.15"
numpy = "^1.26"
pandas = { extras = ["performance", "pyarrow"], version = "^2.2.0" }
flask_sqlalchemy = "^3.1"
sqlalchemy = "^2.0.18"
scipy = "^1.12"
redis = { extras = ["hiredis"], version = "^5.0.1" }
dash-bootstrap-components = "^1.3.0"
flask = "*"
dash_daq = "0.5.0"
plotly = "5.8.2"
colorlover = "0.3.0"
requests = "^2.31"
pytz = "2024.1"
orjson = "^3.9.14"
paramiko = "^3.4"
python-dotenv = "^1.0.1"
upedata = { path = "dependencies/upedata-0.6.2-py3-none-any.whl" }
werkzeug = "^3.0"
python-dateutil = "^2.8.2"
gunicorn = "^21.2"
psycopg = { extras = ["binary"], version = "^3.1.14" }
pyarrow = "^15.0.0"

[tool.poetry.group.telemetry.dependencies]
azure-monitor-opentelemetry = "^1.4.2"
opentelemetry-instrumentation-flask = "^0.45b0"
opentelemetry-instrumentation-redis = "^0.45b0"
opentelemetry-instrumentation-sqlalchemy = "^0.45b0"
opentelemetry-instrumentation-system-metrics = "^0.45b0"
opentelemetry-instrumentation-logging = "^0.45b0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
coverage = "^7.3.2"
pytest-xdist = "^3.3.1"
psutil = "*"
multiprocess = "*"
selenium = "*"
icecream = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"